2010-03-21  Tatsuhiro Tsujikawa  <t-tujikawa@users.sourceforge.net>
	Defined gid_t as int64_t.
	* src/BtDependency.cc
	* src/BtStopDownloadCommand.cc
	* src/FtpNegotiationCommand.cc
	* src/HttpResponseCommand.cc
	* src/RequestGroup.cc
	* src/RequestGroup.h
	* src/RequestGroupMan.cc
	* src/TrackerWatcherCommand.cc
	* src/XmlRpcMethodImpl.cc
	* src/message.h
	* test/XmlRpcMethodTest.cc

+++ src/RequestGroup.h
@@ -73,7 +73,7 @@ class BtRuntime;
 class PeerStorage;
 #endif // ENABLE_BITTORRENT
 
-typedef int32_t gid_t;
+typedef int64_t gid_t;
 
 class RequestGroup {
 public:
